requirements.txt
asgiref==3.4.1
boto3==1.18.11
botocore==1.21.11
certifi==2021.5.30
Django==3.2.5
django-admin-sortable==2.2.4
django-cleanup==5.2.0
django-storages==1.11.1
django-tinymce==3.3.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6

django-heroku